refactor: getStations 로직 Line으로 이동
